changed long to int

+++ src/TourManager/TourManager.Repository.Abstraction/Entities/Booking.cs
@@ -2,9 +2,9 @@
 {
     public class Booking
     {
-        public long Id { get; set; }
-        public long ExternalId { get; set; }
-        public long TourId { get; set; }
+        public int Id { get; set; }
+        public int ExternalId { get; set; }
+        public int TourId { get; set; }
         public Tour Tour { get; set; }
     }
 }
