+++ src/TourManager/TourManager.Repository.Abstraction/Entities/Client.cs
@@ -4,8 +4,8 @@ namespace TourManager.Repository.Entities
 {
     public class Client
     {
-        public long Id { get; set; }
-        public long ExternalId { get; set; }
+        public int Id { get; set; }
+        public int ExternalId { get; set; }
         public ICollection<TourClient> TourClients { get; set; }
     }
 }
