+++ src/TourManager/TourManager.Repository.Abstraction/Entities/TourClient.cs
@@ -2,10 +2,10 @@
 {
     public class TourClient
     {
-        public long Id { get; set; }
-        public long TourId { get; set; }
+        public int Id { get; set; }
+        public int TourId { get; set; }
         public Tour Tour { get; set; }
-        public long ClientId { get; set; }
+        public int ClientId { get; set; }
         public Client Client { get; set; }
     }
 }
